gettext: cleanup + update fr

+++ src/installer/po/install_po.sh
@@ -1,36 +1,29 @@
-#!/bin/sh
+#!/bin/bash
 # This file is under LGPL license
-# March 2007, Gilles Casse <gcasse@oralux.org>
+# 2007-2019, Gilles Casse <gcasse@oralux.org>
 #
 # install_po.sh : install the newly translated messages
 #
 
+lang="$1"
+default=${LANG:0:2}
+[ -z "$lang" ] && read -p "Please, enter your language code (default = $default): " lang
+[ -z "$lang" ] && lang=$default
+[ "$lang" = en ] && exit 0
 
-if [ "$#" != "1" ]; then
-    echo "Please, enter your language code (default = $LANG)"
-    read lg
-else
-    lg="$1"
-fi
 
-# if lg is empty, $LANG is used 
+# The directory locale/$lang/LC_MESSAGES is created if it does not exist
 
-case v$lg in
-  v) lg=$LANG;;
-esac
-
-# The directory locale/$lg/LC_MESSAGES is created if it does not exist
-
-LOCALE="$(dirname $(pwd))/locale/$lg/LC_MESSAGES"
+LOCALE="$(dirname $(pwd))/locale/$lang/LC_MESSAGES"
 
 install -d $LOCALE
 
-if [ -e "$lg.new.po" ]; then 
-    mv $lg.def.po $lg.def.po.bak
-    mv $lg.new.po $lg.def.po
+if [ -e "$lang.new.po" ]; then 
+    mv $lang.def.po $lang.def.po.bak
+    mv $lang.new.po $lang.def.po
 fi
 
-msgfmt $lg.def.po -o voxin.mo
+msgfmt $lang.def.po -o voxin.mo
 mv voxin.mo $LOCALE
 
 echo "...the file voxin.mo is installed in $LOCALE"
